fix: downgrade u128 to u64 and handle integer overflow
* fix: wrap fund account with box

+++ programs/restaking/src/fund/events/info.rs
@@ -7,8 +7,8 @@ pub struct FundInfo {
     pub admin: Pubkey,
     pub lrt_mint: Pubkey,
     pub supported_tokens: Vec<TokenInfo>,
-    pub sol_amount_in: u128,
-    pub sol_reserved_amount: u128,
+    pub sol_amount_in: u64,
+    pub sol_reserved_amount: u64,
     pub sol_withdrawal_fee_rate: f32,
     pub sol_withdrawal_enabled: bool,
 }
